+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ## 0.5.46 (Pending):
+* Add CollectorEntityChangeListener that records entity change events, useful during testing.
 * Introduce ReplicantClientTestModule in the client-side test code that registers the standard set
   of client-side dependencies. Update AbstractClientTest to add ReplicantClientTestModule to the
   list of modules.